copywriter/requirements.txt
streamlit>=1.28.0
openai>=1.0.0
python-docx>=0.8.11
pdfplumber>=0.10.0
python-dotenv>=1.0.0
supabase>=2.0.0
